# components/backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/card/issue": {
      "post": {
        "description": "Выпуск карты",
        "operationId": "issueCard",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueCardInputDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешный выпуск карты",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueCardResponseDTO"
                }
              }
            }
          }
        },
        "summary": "issueCard",
        "tags": [
          "Card"
        ]
      }
    },
...
    "/card/{card_id}/private-data": {
      "get": {
        "description": "Получение приватных данных по ID карты",
        "operationId": "getPrivateData",
        "parameters": [
          {
            "name": "card_id",
            "required": true,
            "in": "path",
            "description": "ID карты",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Приватные данные пользователя",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivateDataResponseDTO"
                }
              }
            }
          }
        },
        "summary": "getPrivateData",
        "tags": [
          "Card"
        ]
      }
    },
    "/card/{card_id}": {
      "delete": {
        "description": "Деактивация карты",
        "operationId": "deactivateCard",
        "parameters": [
          {
            "name": "card_id",
            "required": true,
            "in": "path",
            "description": "ID карты для деактивации",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Карта успешно деактивирована",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueCardResponseDTO"
                }
              }
            }
          }
        },
        "summary": "deactivateCard",
        "tags": [
          "Card"
        ]
      }
    },
...
  },
  "info": {
    "title": "CARD.COOP",
    "description": "API для авторизации и управления пользователями",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3015"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CardMetaDTO": {
        "type": "object",
        "properties": {
          "version": {
            "type": "number",
            "description": "Версия модели данных карты",
            "example": 1
          },
          "card_type": {
            "type": "string",
            "description": "Тип карты",
            "example": "standard"
          },
          "valid_until": {
            "format": "date-time",
            "type": "string",
            "description": "Дата окончания действия карты",
            "example": "2030-01-01"
          },
          "is_active": {
            "type": "boolean",
            "description": "Активна ли карта",
            "example": true,
            "default": true
          }
        },
        "required": [
          "version",
          "is_active"
        ]
      },
      "IssueCardInputDTO": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "Имя пользователя, которому будет выдана карта",
            "example": "user123"
          },
          "coopname": {
            "type": "string",
            "description": "Название кооператива",
            "example": "coop_example"
          },
          "encrypted_key": {
            "type": "string",
            "description": "Зашифрованный приватный ключ",
            "example": "encryptedKey123456abcdef"
          },
          "encrypted_data": {
            "type": "string",
            "description": "Зашифрованные данные, связанные с картой",
            "example": "encryptedData123456abcdef"
          },
          "meta": {
            "description": "Метаданные, связанные с картой",
            "example": {
              "version": 1,
              "card_type": "standard"
            },
            "allOf": [
              {
                "$ref": "#/components/schemas/CardMetaDTO"
              }
            ]
          }
        },
        "required": [
          "username",
          "coopname",
          "encrypted_key",
          "encrypted_data",
          "meta"
        ]
      },
      "IssueCardResponseDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Уникальный идентификатор карты",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "username": {
            "type": "string",
            "description": "Имя пользователя, связанного с картой",
            "example": "username123"
          },
          "coopname": {
            "type": "string",
            "description": "Название кооператива",
            "example": "coop_example"
          },
          "issued_at": {
            "format": "date-time",
            "type": "string",
            "description": "Дата выпуска карты",
            "example": "2023-01-01T12:00:00Z"
          },
          "card_type": {
            "type": "string",
            "description": "Тип карты",
            "example": "standard"
          },
          "is_active": {
            "type": "boolean",
            "description": "Активна ли карта",
            "example": true
          }
        },
        "required": [
          "id",
          "username",
          "coopname",
          "issued_at",
          "card_type",
          "is_active"
        ]
      },
...
      "PrivateDataResponseDTO": {
        "type": "object",
        "properties": {
          "encrypted_data": {
            "type": "string",
            "description": "Зашифрованные данные пользователя",
            "example": "encrypted-data-string"
          },
          "data_hash": {
            "type": "string",
            "description": "Хеш данных для верификации",
            "example": "data-hash-string"
          },
          "version": {
            "type": "number",
            "description": "Версия формата данных",
            "example": 1
          },
          "data_type": {
            "type": "string",
            "description": "Тип приватных данных",
            "example": "personal"
          }
        },
        "required": [
          "encrypted_data",
          "data_hash",
          "version",
          "data_type"
        ]
      },
...
    }
  }
}